+++ server/routes/index.js
@@ -1,0 +1,13 @@
+/**
+ * @file Routing indices.
+ *
+ * @author Koichi Nagaoka
+ */
+
+const router = require('./router');
+
+//
+// TODO : Define routing rules.
+//
+
+module.exports = router;