+++ server/routes/router.js
@@ -1,0 +1,10 @@
+/**
+ * @file Routing instance.
+ *
+ * @author Koichi Nagaoka
+ */
+
+const express     = require('express');
+const router      = express.Router();
+
+module.exports = router;